exporter: allow attestation options to contain list of reasons
Instead of just the boolean true/false values, we allow the attestation
option for exporters to contain an arbitrary list of "attestation
reasons". Only attestations that have a reason matching the list will
actually be output.

This allows clients to completely detach the concepts of "what
attestations to generate" and "what attestations to output".

Signed-off-by: Justin Chadwell <me@jedevc.com>