dockerfile: collect all dependencies as extra scan targets
Previously, we only checked for direct dependencies of the stage, and
direct dependencies of any of it's parents. However, transitive
dependencies would not be detected.

To resolve this, we add a new findReachable function, to find all
dependencies, including the transitive dependencies. The logic of this
new function doesn't need to worry about circular dependencies: those
should never be returned from the toDispatchState function as they are
invalid.

Signed-off-by: Justin Chadwell <me@jedevc.com>