ociindex: refactor to hide implementation internally
This allows some of the client processing code to abstract more over the
details of how the ociindex package works, and how it loads tags - the
calling code only need to know that it's being pointed at a containerd
content store directory.

Importantly, this allows for reusability, so we can use this same code
more easily in buildx for oci-layouts.

Signed-off-by: Justin Chadwell <me@jedevc.com>