Add get-user-info utility
This utility allows us to mount the buildkitd executable inside a
Windows container and fetch the SID of any existing user. This is to
work around the fact that the SAM hive data structures are undocumented
and there is no API to inspect an offline SAM hive to fetch the security
info of an existing user.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>