dockerfile: allow ARGs in COPY --from and RUN --mount=from
Previously, using ARGs in COPY --from or RUN --mounts from was not
supported, requiring using an extra intermediate stage.  This patch
allows global args to be used directly in these locations.

To do this, we need to resolve a dependency issue:
- We need to construct all the dispatchStates before evaluating.
- To construct the dispatchStates, we need to determine the
  fully-evaluated base LLB source.
- However, to fully-evaluate the base name using the Expand interface,
  we need to have constructed all the dispatchStates.

It is theoretically possible to refactor to not have this restriction,
however, this is not required to implement the base functionality for
global args.

To resolve this, we can use just the global args to evaluate the FROM
fields using the toCommand helper, where the stage resolution occurs
today.

Signed-off-by: Justin Chadwell <me@jedevc.com>