buildctl: Provide --wait option
Added `--wait` to buildctl's global options. See below for behavior.

Implemented a `Wait` client method that blocks until a successful
request has been made to the remote buildkit. This behavior is identical
as in buildx, and only makes additional ListWorker calls *if the user
has requested them*.

The timeout as requested using the `--timeout` option is additionally
applied here.

Co-authored-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: Justin Chadwell <me@jedevc.com>