test: when content still exist, prune again
Sometimes, a saved trace is still in the content store and is not
removed by garbage collection call due to timing issues (a history trace
is saved in containerd while it is running the garbage collection).

When this happens, trigger a garbage collection again so the remaining
content is removed.

Signed-off-by: Alexis Murzeau <amubtdx@gmail.com>