Add path handling functions
There are several places throughout the code where we repeat the same
normalization patterns for paths. This change adds a couple of helper
functions and makes CheckSystemDriveAndRemoveDriveLetter() on all
platforms.

Additionally, all functions accept a platform flag that indicates the
target OS we're building the image for. Decissions are made based on
this flag instead of checking the platform on which buildkit is running.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>