hack: forward CGO_ENABLED into test container
If we try and run the integration tests with `-race` set, `hack/test`
correctly prompts us to also set `CGO_ENABLED`.

However, this value only gets set during build, and not during the
actual test run, which spawns the following error:

	=== Errors
	go: -race requires cgo; enable cgo by setting CGO_ENABLED=1

	DONE 0 tests, 1 error in 0.002s

To resolve this, we need to propogate the value of `CGO_ENABLED` into
the created test containers, so that we can actually run these tests
properly.

Signed-off-by: Justin Chadwell <me@jedevc.com>