progressui: json mode for a buffered json output
This adds a json progress mode to the progressui. This json mode, in
contrast to `rawjson`, produces an aggregated json output for each
vertex after the vertex has completed.

The primary use of this progress output is for machine tools to parse
to determine the details of the build. For raw events and interactive
tooling, `rawjson` is a better option due to providing more real time
detail.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>