refactor(cache): make impls of RefOption public
Previously, RefOption usage was a mix of public and
private types behind an empty interface.

It was not possible to implement the `Accessor` interface as
many RefOptions types were private.

Now, using a private interface the list of options types is fixed.
The options are now public so `Accessors` can be implemented.

Signed-off-by: Chris Goller <goller@gmail.com>