Fix ResolveImageConfig to evaluate source policy
Backport 330cf7ae7d9ac983432ab13f8bc9e7d4aff75695 to 0.11 branch

This is not a clean cherry-pick due to lots of other interspersed
changes (and this change really has to touch code all up and down the
stack).

Original commit message:

Before this change, ResolveImageConfig was unaware of source policies.
This means that:

1. Images for denied sources may be resolved
2. Image configs may get pulled for sources that are later converted to
   a different image

The update makes it so the image resolver first runs a given ref through
the source policy and uses any mutated ref for the actual resolve
(instead of the original ref).
It also returns the mutated ref so it can be used correctly by the
frontend (e.g. don't want to do llb.Image(oldRef@resolvedDigest)).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>