Fix source police mutation for named context
This was unfortuantely just wrong.
The reference for a named context is is fetched using
`nameWithPlatform`.
When a source policy mutated the ref this was recursively trying to
resolve the reference again, however it was using the mutated ref as the
new context name (and even that was not correct because `name` was made
the new `nameWithPlatform` which is what is actually used as the key
rather than `name`).

This shouldn't be mutating the context key but rather the ref the
context is pointing to.

This change updates the build opts such the value pointed at by
`nameWithPlatform` in the build opts map points to the updated ref
instead of the original.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>