Dockerfile: add ALPINE_CACHE image to lock apk versions
The `ALPINE_CACHE` image contains `/etc/apk/cache` for apk files
referred from the entire Dockerfile.

All the `apk add` commands in the Dockerfile are now executed with
the mount for `ALPINE_CACHE` and with `--no-network`.

The `ALPINE_CACHE` image defaults to the `alpine-cache-local` stage,
but can be also set to `docker.io/moby/buildkit:${RELEASE}-alpine-cache`
for reproducibility.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>