🚨WIP: OCI Images used as cache SHOULD not have empty layers list
Per image-spec 1.1.0, OCI Images that would otherwise have an empty
layers list SHOULD instead have a single "Empty JSON" known-value layer
listed.

This guideline does not apply to Image Indexes, and also does not appear
in the Docker Manifest/Manifest List documentation, so those cases are
unchanged.

WIP notes:

The empty layer is not handled in the cache importer. Not sure if we
need to, tests pass without any such changes.

Do we actually need to ensure the blob exists on the registry? Probably;
the tests pass because the blob already exists as the default cache
config object, but that's 100% crimes.

Only applied to remote registry cache. Not sure if the other caches
should do something similar. Are any of them also outputting in an OCI
image format?

Relatedly, should we do something like this for normal OCI images? I
don't know how robust the ecosystem is against unexpected layers in an
OCI Image.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>