vendor: upgrade OpenTelemetry libraries to v1.19.0 / v0.45.0
Core OTel libs are now at v1.19.0, while contrib is at v0.45.0.
The protocol has also reached 1.0.0.

Note that Jaeger exporter is frozen at v1.17.0, which is the
last version it was supported by OpenTelemetry. BuildKit should
likely begin the process of deprecating it as well, since OTLP
is now the recommended solution (by Jaeger themselves), but that
is not handled here.

Co-authored-by: CrazyMax <crazy-max@users.noreply.github.com>
Signed-off-by: CrazyMax <crazy-max@users.noreply.github.com>